downgrade to jre 1.4

+++ web/WEB-INF/classes/src/oscar/oscarEncounter/oscarConsultationRequest/pageUtil/EctConsultationFormRequestPrintPdf.java
@@ -175,7 +175,7 @@ public class EctConsultationFormRequestPrintPdf {
                 dynamicHeight = LINEHEIGHT - 152;
             }
             
-            ct.setSimpleColumn(new Float(85), height - 264 - dynamicHeight - lineCount*LINEHEIGHT, new Float(526), height - 250 - dynamicHeight, LINEHEIGHT, Element.ALIGN_LEFT);
+            ct.setSimpleColumn(new Float(85).floatValue(), height - 264 - dynamicHeight - lineCount*LINEHEIGHT, new Float(526).floatValue(), height - 250 - dynamicHeight, LINEHEIGHT, Element.ALIGN_LEFT);
             ct.addText(new Phrase(name, boldFont));
             ct.addText(new Phrase(text, font));
             ct.go();
@@ -200,7 +200,7 @@ public class EctConsultationFormRequestPrintPdf {
         cb.showTextAligned(PdfContentByte.ALIGN_LEFT, reqForm.specPhone, 190, height - 166, 0);
         cb.showTextAligned(PdfContentByte.ALIGN_LEFT, reqForm.specFax, 190, height - 181, 0);
         cb.endText();
-        ct.setSimpleColumn(new Float(190), height - 223, new Float(290), height - 181, LINEHEIGHT, Element.ALIGN_LEFT);
+        ct.setSimpleColumn(new Float(190).floatValue(), height - 223, new Float(290).floatValue(), height - 181, LINEHEIGHT, Element.ALIGN_LEFT);
         ct.addText(new Phrase(reqForm.specAddr.replaceAll("<br>", "\n"), font));
         ct.go();
         
@@ -209,7 +209,7 @@ public class EctConsultationFormRequestPrintPdf {
         cb.setFontAndSize(bf, FONTSIZE);
         cb.showTextAligned(PdfContentByte.ALIGN_LEFT, reqForm.patientName, 385, height - 112, 0);
         cb.endText();
-        ct.setSimpleColumn(new Float(385), height - 153, new Float(585), height - 112, LINEHEIGHT, Element.ALIGN_LEFT);
+        ct.setSimpleColumn(new Float(385).floatValue(), height - 153, new Float(585).floatValue(), height - 112, LINEHEIGHT, Element.ALIGN_LEFT);
         ct.addText(new Phrase(reqForm.patientAddress.replaceAll("<br>", "\n"), font));
         ct.go();
